+++ test/phoenix/router/route_test.exs
@@ -4,7 +4,7 @@ defmodule Phoenix.Router.RouteTest do
   import Phoenix.Router.Route
 
   test "builds a route based on verb, path, controller, action and helper" do
-    route = build("GET", "/foo/:bar", Hello, :world, "hello_world", [:foo, :bar])
+    route = build("GET", "/foo/:bar", Hello, :world, "hello_world", [:foo, :bar], "example.com")
     assert route.verb == "GET"
     assert route.segments == ["foo", {:bar, [], nil}]
     assert route.binding == [{"bar", {:bar, [], nil}}]